fix: move ContentTypesAdminNode format migration to OrchardCore.AdminMenu
